fix: reuse step helper text in output variables tab
