+++ source/javascripts/components/StepConfigPanel/StepConfiguration.tsx
