+++ source/javascripts/components/StepConfigPanel/StepProperties.tsx
