+++ source/javascripts/components/StepConfigPanel/components/StepInputList.tsx
