+++ source/javascripts/controllers/_WorkflowsController.js.erb
@@ -1,7 +1,7 @@
 import { safeDigest } from "../services/react-compat";
 import datadogRumCustomTiming from "../utils/datadogCustomRumTiming.ts";
 import { getVersionRemark } from "../utils/stepVersionUtil";
-import { extractInputNames, extractReleaseNotesUrl } from "../components/StepProperties/utils";
+import { extractInputNames, extractReleaseNotesUrl } from "../components/StepConfigPanel/utils";
 
 (function() {
   "use strict";
